// types/profilemusic.ts
export interface ProfileMusic {
    id: string;
    profile_id: string;
    deezer_id: string;
    title: string;
    artist: string;
    album: string;
    url: string;
    cover_url: string;
    duration: number;
    preview_url: string;
    preview_expires_at: string;
    created_at: string;
    updated_at: string;
}